requirements.txt
fastapi
uvicorn
python-jose
passlib
bcrypt==3.2.0
strawberry-graphql==0.246.0
pydantic
sqlalchemy